requirements.txt
click==8.1.3
jinja2==3.1.2
py-solc-x==1.1.1
pyyaml==6.0
tomlkit==0.11.6
web3==6.0.0
